集成环信
由于 .a 文件太大, 没有上传

+++ ChuanSongMen/BaseViewController.h
@@ -23,7 +23,7 @@
 @property (strong, nonatomic) AppDelegate *app;
 @property (strong, nonatomic) UILabel *titleLable;
 @property (nonatomic, strong) MBProgressHUD  *hud;
-@property (nonatomic,assign) int userId;
+@property (nonatomic,assign) NSInteger userId;
 
 @property (nonatomic, strong) UserInfoModel *userInfoModel;
